i3: add a wrapper which handles different DPIs
This is kind of cool. It uses the mouse pointer to figure out which
screen you are at, then scales based on that. It could scale dynamically
based on current DPI and resolution, but this is good enough for now.

Also switch the monitors since after the reinstall the ids changed